ADD in src/http/HttpRequest many methods (appendBody, etc.)

+++ include/http/RequestParser.hpp
@@ -31,7 +31,7 @@ static std::string extractMethod(std::istringstream & iss);
 static std::string extractUri(std::istringstream & iss);
 static std::string extractVersion(std::istringstream & iss);
 static std::map<std::string, std::string> extractHeaders(std::string & buffer);
-static int calculateContentLength(const std::map<std::string, std::string> & headers);
+static size_t calculateContentLength(const std::map<std::string, std::string> & headers);
 static std::string extractBody(std::string & buffer, int contentLength);
 
 #endif
